Update to latest version of `csv` package
The `csv` package has split up into separate packages, so we only have
to depend on the parser. The plugin as completely rewritten to use the
new API. All options are passed directly to the parser too, which makes
it a lot easier to configure the parser.

Along the way we swtiched from the `map-stream` to the Gulp "standard"
`through2` package.